deleting a blog is possible only if the token sent with the request is the same as that of the blog's creator

+++ bloglist/controllers/blogs.js
@@ -47,11 +47,36 @@ blogsRouter.get('/', async (request, response) => {
   })
 
 
-  blogsRouter.delete('/:id', async (request,response,next) => {
-
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
-  })
+  blogsRouter.delete('/:id', async (request, response, next) => {
+    try {
+      const decodedToken = jwt.verify(request.token, process.env.SECRET);
+      if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' });
+      }
+  
+      const blog = await Blog.findById(request.params.id);
+      if (!blog) {
+        return response.status(404).json({ error: 'blog not found' });
+      }
+  
+      if (blog.user.toString() !== decodedToken.id.toString()) {
+        return response.status(403).json({ error: 'only the creator can delete this blog' });
+      }
+  
+      // Delete the blog
+      await Blog.findByIdAndDelete(request.params.id);
+  
+      // Find the user and remove the blog from their blogs array
+      const user = await User.findById(decodedToken.id);
+      user.blogs = user.blogs.filter(blogId => blogId.toString() !== request.params.id);
+      await user.save();
+  
+      response.status(204).end();
+    } catch (exception) {
+      next(exception);
+    }
+  });
+  
 
   blogsRouter.put('/:id', async (request,response,next) => {
     const body = request.body  //new content
